Add WS command number/device_class_convertible_units
